Make GlideRecord generic and remove TypedGR
