+++ util/index.d.ts
@@ -1,5 +1,3 @@
 import { Override } from './helpers';
-import { TypedGR } from './TypedGR';
-import { ReferenceGlideElement, TypedRESTAPIRequest } from './strongTypes';
 
-export { TypedGR, ReferenceGlideElement, TypedRESTAPIRequest };
+export { Override };
